Bugfix + Permissions checks
I relaxed the restrictions on properties ever so slightly. The System Object has access to write and read any properties.

I've added a new ScriptGlobals class called AdminScriptGlobals which inherits from ScriptGlobals. It's empty at the moment but the ScriptEngine will automatically load it instead of ScriptGlobals if ThisObject has an Admin flag.

Fixed a bug where SOMETIMES objects were accessed from the Db and not the memory cache. All instances of that should now be fixed.